Login, registration, email verification seem to work
Next up: cleanup and more extensive testing
